sql/: reorganize, make idempotent, generate fresh/
This is a large reorganization of the sql/ files. Now there is only a
single command in each file, so we can directly curs.execute() the
contents through psycopg2. Each file should be idempotent, so we can
blindly execute it against a new or existing db.

The sql/fresh/ symlinks have been redone and are now automatically
regenerated when running schema.py.

schema.py now has an --init argument which will execute each script in
order against the configured database.